component conditions integrated into seignup. Form validation (server) bug fix showing form error in popup when form does not validate. New metadata plugin to support multiline template components inside {} Tabs, new lines, etc.. can be included to make components more readable.
Facebook integration, rerender from external renders to activate components related to external renders or binding actions in external render javascript code. Example, in facebook signup render, we set variable socialLogged to true. We have condition that when true, captcha code is not neccessary. Therefore, we need to re-render conditions and other components to update.

Method ximpia.common.Broser.putAttr() to include variable into server response saved into sessioin storage object. This allows visual javascript code to write to response so that conditions can be re-evaluated.

+++ ximpia/site/models.py
@@ -363,7 +363,7 @@ class SocialNetworkUser ( BaseModel ):
 				verbose_name = _('User'), help_text = _('User'))
 	socialNetwork = models.ForeignKey('core.CoreParam', limit_choices_to={'mode': CoreK.NET}, db_column='ID_CORE_PARAMETER',
 				verbose_name = _('Social Network'), help_text = _('Social network'))
-	socialId = models.IntegerField(db_column='SOCIAL_ID', verbose_name = _('Social ID'), help_text = _('Social network user id'))
+	socialId = models.BigIntegerField(db_column='SOCIAL_ID', verbose_name = _('Social ID'), help_text = _('Social network user id'))
 	token = models.CharField(max_length=255, db_column='TOKEN',
 				verbose_name = _('Token'), help_text = _('Token'))
 	tokenSecret = models.CharField(max_length=255, null=True, blank=True, db_column='TOKEN_SECRET',
